Stage #2: Audio devices and their integration
A module for detecting and outputting all audio devices on the machine was added.
Module processing for detecting and displaying all audio devices has been added
A module was added to test connection to audio devices
Module processing for connecting to an audio device has been added
A module for capturing audio streams from audio devices was added
Processing of the module for capturing audio streams was added
Module for outputting audio streams to file and audio device was added
Processing of audio stream output module was added
Added logging and docstring for all modules
Added testing of modules for operation and correctness of execution.
Display of all audio devices in run.md output.